Added spinners to clients and updated scheduling
Client entry now uses spinners for numeric values (besides phone
number). When scheduling clients all the instructors now get a text and
an invite to the event.

Fixed a bug where deleting an instructor would take the user to the
client screen.

Need to finish text information to instructors and check for empty or
incorrect phone numbers.